+++ includes/class-comment.php
@@ -19,10 +19,15 @@ class Comment {
 	 * Initialize the class, registering WordPress hooks
 	 */
 	public static function init() {
+		self::register_comment_types();
+
 		\add_filter( 'comment_reply_link', array( self::class, 'comment_reply_link' ), 10, 3 );
 		\add_filter( 'comment_class', array( self::class, 'comment_class' ), 10, 3 );
 		\add_filter( 'get_comment_link', array( self::class, 'remote_comment_link' ), 11, 3 );
 		\add_action( 'wp_enqueue_scripts', array( self::class, 'enqueue_scripts' ) );
+		\add_action( 'pre_get_comments', array( static::class, 'comment_query' ) );
+
+		\add_filter( 'get_avatar_comment_types', array( static::class, 'get_avatar_comment_types' ), 99 );
 	}
 
 	/**
@@ -462,5 +467,160 @@ class Comment {
 				$assets['version']
 			);
 		}
+	}
+
+	/**
+	 * Return the registered custom comment types.
+	 *
+	 * @return array The registered custom comment types
+	 */
+	public static function get_comment_types() {
+		global $activitypub_comment_types;
+
+		return $activitypub_comment_types;
+	}
+
+	/**
+	 * Is this a registered comment type
+	 *
+	 * @param string $slug The name of the type
+	 * @return boolean True if registered.
+	 */
+	public static function is_registered_comment_type( $slug ) {
+		$slug = strtolower( $slug );
+		$slug = sanitize_key( $slug );
+
+		return in_array( $slug, array_keys( self::get_comment_types() ), true );
+	}
+
+	/**
+	 * Return the registered custom comment types names.
+	 *
+	 * @return array The registered custom comment type names
+	 */
+	public static function get_comment_type_names() {
+		return array_values( wp_list_pluck( self::get_comment_types(), 'type' ) );
+	}
+
+	/**
+	 * Get a comment type
+	 *
+	 * @param string $type The comment type
+	 *
+	 * @return array The comment type
+	 */
+	public static function get_comment_type( $type ) {
+		$type  = strtolower( $type );
+		$type  = sanitize_key( $type );
+		$types = self::get_comment_types();
+
+		if ( in_array( $type, array_keys( $types ), true ) ) {
+			$type_array = $types[ $type ];
+		} else {
+			$type_array = array();
+		}
+
+		return apply_filters( "activitypub_comment_type_{$type}", $type_array );
+	}
+
+	/**
+	 * Get a comment type attribute
+	 *
+	 * @param string $type The comment type
+	 * @param string $attr The attribute to get
+	 *
+	 * @return mixed The value of the attribute
+	 */
+	public static function get_comment_type_attr( $type, $attr ) {
+		$type_array = self::get_comment_type( $type );
+
+		if ( $type_array && isset( $type_array[ $attr ] ) ) {
+			$value = $type_array[ $attr ];
+		} else {
+			$value = '';
+		}
+
+		return apply_filters( "activitypub_comment_type_{$attr}", $value, $type );
+	}
+
+
+
+	/**
+	 * Register the comment types used by the ActivityPub plugin
+	 *
+	 * @return void
+	 */
+	public static function register_comment_types() {
+		register_comment_type(
+			'announce',
+			array(
+				'label'       => __( 'Reposts', 'activitypub' ),
+				'singular'    => __( 'Repost', 'activitypub' ),
+				'description' => __( 'A repost on the indieweb is a post that is purely a 100% re-publication of another (typically someone else\'s) post.', 'activitypub' ),
+				'icon'        => '♻️',
+				'class'       => 'p-repost',
+				'type'        => 'repost',
+				// translators: %1$s username, %2$s opject format (post, audio, ...), %3$s URL, %4$s domain
+				'excerpt'     => __( '&hellip; reposted this!', 'activitypub' ),
+			)
+		);
+
+		register_comment_type(
+			'like',
+			array(
+				'label'       => __( 'Likes', 'activitypub' ),
+				'singular'    => __( 'Like', 'activitypub' ),
+				'description' => __( 'A like is a popular webaction button and in some cases post type on various silos such as Facebook and Instagram.', 'activitypub' ),
+				'icon'        => '👍',
+				'class'       => 'p-like',
+				'type'        => 'like',
+				// translators: %1$s username, %2$s opject format (post, audio, ...), %3$s URL, %4$s domain
+				'excerpt'     => __( '&hellip; liked this!', 'activitypub' ),
+			)
+		);
+	}
+
+	/**
+	 * Show avatars on Activities if set
+	 *
+	 * @param array $types list of avatar enabled comment types
+	 *
+	 * @return array show avatars on Activities
+	 */
+	public static function get_avatar_comment_types( $types ) {
+		$comment_types = self::get_comment_type_names();
+		$types         = array_merge( $types, $comment_types );
+
+		return array_unique( $types );
+	}
+
+	/**
+	 * Excludes likes and reposts from comment queries.
+	 *
+	 * @author Jan Boddez
+	 *
+	 * @see https://github.com/janboddez/indieblocks/blob/a2d59de358031056a649ee47a1332ce9e39d4ce2/includes/functions.php#L423-L432
+	 *
+	 * @param  WP_Comment_Query $query Comment count.
+	 */
+	public static function comment_query( $query ) {
+		if ( ! $query instanceof WP_Comment_Query ) {
+			return;
+		}
+
+		if ( is_admin() || ! is_singular() ) {
+			return;
+		}
+
+		if ( ! empty( $query->query_vars['type__in'] ) ) {
+			return;
+		}
+
+		if ( isset( $query->query_vars['count'] ) && true === $query->query_vars['count'] ) {
+			return;
+		}
+
+		// Exclude likes and reposts by the Webmention plugin.
+		$query->query_vars['type__not_in'] = self::get_comment_type_names();
 	}
 }