+++ includes/functions.php
@@ -9,6 +9,7 @@ use Activitypub\Webfinger;
 use Activitypub\Activity\Activity;
 use Activitypub\Collection\Followers;
 use Activitypub\Collection\Users;
+use Activitypub\Collection\Extra_Fields;
 
 /**
  * Returns the ActivityPub default JSON-context
@@ -209,8 +210,7 @@ function is_comment() {
 	if ( ! is_null( $comment_id ) ) {
 		$comment = \get_comment( $comment_id );
 
-		// Only return local origin comments
-		if ( $comment && $comment->user_id ) {
+		if ( $comment ) {
 			return $comment_id;
 		}
 	}
@@ -656,11 +656,11 @@ function get_active_users( $duration = 1 ) {
 
 	// if blog user is disabled
 	if ( is_user_disabled( Users::BLOG_USER_ID ) ) {
-		return $count;
+		return (int) $count;
 	}
 
 	// also count blog user
-	return $count + 1;
+	return (int) $count + 1;
 }
 
 /**
@@ -688,10 +688,10 @@ function get_total_users() {
 
 	// if blog user is disabled
 	if ( is_user_disabled( Users::BLOG_USER_ID ) ) {
-		return $users;
+		return (int) $users;
 	}
 
-	return $users + 1;
+	return (int) $users + 1;
 }
 
 /**
@@ -742,8 +742,13 @@ function object_to_uri( $object ) { // phpcs:ignore Universal.NamingConventions.
 		return $object;
 	}
 
+	$type = 'Object';
+	if ( isset( $object['type'] ) ) {
+		$type = $object['type'];
+	}
+
 	// return part of Object that makes most sense
-	switch ( $object['type'] ) {
+	switch ( $type ) {
 		case 'Link':
 			$object = $object['href'];
 			break;
@@ -1012,6 +1017,39 @@ function custom_large_numbers( $formatted, $number, $decimals ) {
 }
 
 /**
+ * Registers a ActivityPub comment type.
+ *
+ *
+ * @param string $comment_type Key for comment type.
+ * @param array  $args         Arguments.
+ *
+ * @return array The registered Activitypub comment type.
+ */
+function register_comment_type( $comment_type, $args = array() ) {
+	global $activitypub_comment_types;
+
+	if ( ! is_array( $activitypub_comment_types ) ) {
+		$activitypub_comment_types = array();
+	}
+
+	// Sanitize comment type name.
+	$comment_type = sanitize_key( $comment_type );
+
+	$activitypub_comment_types[ $comment_type ] = $args;
+
+	/**
+	 * Fires after a ActivityPub comment type is registered.
+	 *
+	 *
+	 * @param string $comment_type Comment type.
+	 * @param array  $args         Arguments used to register the comment type.
+	 */
+	do_action( 'activitypub_registered_comment_type', $comment_type, $args );
+
+	return $args;
+}
+
+/**
  * Normalize a URL.
  *
  * @param string $url The URL.
@@ -1039,34 +1077,6 @@ function normalize_host( $host ) {
 }
 
 /**
- * Get the Extra Fields of an Actor
- *
- * @param int $user_id The User-ID.
- *
- * @return array The extra fields.
- */
-function get_actor_extra_fields( $user_id ) {
-	$extra_fields = new WP_Query(
-		array(
-			'post_type' => 'ap_extrafield',
-			'nopaging'  => true,
-			'status'    => 'publish',
-			'author'    => $user_id,
-			'orderby'   => 'menu_order',
-			'order'     => 'ASC',
-		)
-	);
-
-	if ( $extra_fields->have_posts() ) {
-		$extra_fields = $extra_fields->posts;
-	} else {
-		$extra_fields = array();
-	}
-
-	return apply_filters( 'activitypub_get_actor_extra_fields', $extra_fields, $user_id );
-}
-
-/**
  * Get the reply intent URI.
  *
  * @return string The reply intent URI.
@@ -1076,4 +1086,132 @@ function get_reply_intent_uri() {
 		'javascript:(()=>{window.open(\'%s\'+encodeURIComponent(window.location.href));})();',
 		esc_url( \admin_url( 'post-new.php?in_reply_to=' ) )
 	);
+}
+
+/**
+ * Replace content with links, mentions or hashtags by Regex callback and not affect protected tags.
+ *
+ * @param $content        string   The content that should be changed
+ * @param $regex          string   The regex to use
+ * @param $regex_callback callable Callback for replacement logic
+ *
+ * @return string The content with links, mentions, hashtags, etc.
+ */
+function enrich_content_data( $content, $regex, $regex_callback ) {
+	// small protection against execution timeouts: limit to 1 MB
+	if ( mb_strlen( $content ) > MB_IN_BYTES ) {
+		return $content;
+	}
+	$tag_stack = array();
+	$protected_tags = array(
+		'pre',
+		'code',
+		'textarea',
+		'style',
+		'a',
+	);
+	$content_with_links = '';
+	$in_protected_tag = false;
+	foreach ( wp_html_split( $content ) as $chunk ) {
+		if ( preg_match( '#^<!--[\s\S]*-->$#i', $chunk, $m ) ) {
+			$content_with_links .= $chunk;
+			continue;
+		}
+
+		if ( preg_match( '#^<(/)?([a-z-]+)\b[^>]*>$#i', $chunk, $m ) ) {
+			$tag = strtolower( $m[2] );
+			if ( '/' === $m[1] ) {
+				// Closing tag.
+				$i = array_search( $tag, $tag_stack, true );
+				// We can only remove the tag from the stack if it is in the stack.
+				if ( false !== $i ) {
+					$tag_stack = array_slice( $tag_stack, 0, $i );
+				}
+			} else {
+				// Opening tag, add it to the stack.
+				$tag_stack[] = $tag;
+			}
+
+			// If we're in a protected tag, the tag_stack contains at least one protected tag string.
+			// The protected tag state can only change when we encounter a start or end tag.
+			$in_protected_tag = array_intersect( $tag_stack, $protected_tags );
+
+			// Never inspect tags.
+			$content_with_links .= $chunk;
+			continue;
+		}
+
+		if ( $in_protected_tag ) {
+			// Don't inspect a chunk inside an inspected tag.
+			$content_with_links .= $chunk;
+			continue;
+		}
+
+		// Only reachable when there is no protected tag in the stack.
+		$content_with_links .= \preg_replace_callback( $regex, $regex_callback, $chunk );
+	}
+
+	return $content_with_links;
+}
+
+/**
+ * Generate a summary of a post.
+ *
+ * This function generates a summary of a post by extracting:
+ *
+ * 1. The post excerpt if it exists.
+ * 2. The first part of the post content if it contains the <!--more--> tag.
+ * 3. An excerpt of the post content if it is longer than the specified length.
+ *
+ * @param int|WP_Post $post   The post ID or post object.
+ * @param integer     $length The maximum length of the summary.
+ *                            Default is 500. It will ne ignored if the post excerpt
+ *                            and the content above the <!--more--> tag.
+ *
+ * @return string The generated post summary.
+ */
+function generate_post_summary( $post, $length = 500 ) {
+	$post = get_post( $post );
+
+	if ( ! $post ) {
+		return '';
+	}
+
+	$content = \sanitize_post_field( 'post_excerpt', $post->post_excerpt, $post->ID );
+
+	if ( $content ) {
+		return \apply_filters( 'the_excerpt', $content );
+	}
+
+	$content       = \sanitize_post_field( 'post_content', $post->post_content, $post->ID );
+	$content_parts = \get_extended( $content );
+
+	$excerpt_more = \apply_filters( 'activitypub_excerpt_more', '[…]' );
+	$length       = $length - strlen( $excerpt_more );
+
+	// Check for the <!--more--> tag.
+	if (
+		! empty( $content_parts['extended'] ) &&
+		! empty( $content_parts['main'] )
+	) {
+		$content = $content_parts['main'] . ' ' . $excerpt_more;
+		$length  = null;
+	}
+
+	$content = \html_entity_decode( $content );
+	$content = \wp_strip_all_tags( $content );
+	$content = \trim( $content );
+	$content = \preg_replace( '/\R+/m', "\n\n", $content );
+	$content = \preg_replace( '/[\r\t]/', '', $content );
+
+	if ( $length && \strlen( $content ) > $length ) {
+		$content = \wordwrap( $content, $length, '</activitypub-summary>' );
+		$content = \explode( '</activitypub-summary>', $content, 2 );
+		$content = $content[0] . ' ' . $excerpt_more;
+	}
+
+	/* Removed until this is merged: https://github.com/mastodon/mastodon/pull/28629
+	return \apply_filters( 'the_excerpt', $content );
+	*/
+	return $content;
 }