+++ includes/handler/class-undo.php
@@ -3,6 +3,7 @@ namespace Activitypub\Handler;
 
 use Activitypub\Collection\Users;
 use Activitypub\Collection\Followers;
+use Activitypub\Comment;
 
 use function Activitypub\object_to_uri;
 
@@ -28,10 +29,16 @@ class Undo {
 	 */
 	public static function handle_undo( $activity ) {
 		if (
-			isset( $activity['object']['type'] ) &&
-			'Follow' === $activity['object']['type'] &&
-			isset( $activity['object']['object'] )
+			! isset( $activity['object']['type'] ) ||
+			! isset( $activity['object']['object'] )
 		) {
+			return;
+		}
+
+		$type = $activity['object']['type'];
+
+		// Handle "Unfollow" requests
+		if ( 'Follow' === $type ) {
 			$user_id = object_to_uri( $activity['object']['object'] );
 			$user = Users::get_by_resource( $user_id );
 
@@ -45,6 +52,24 @@ class Undo {
 			$actor   = object_to_uri( $activity['actor'] );
 
 			Followers::remove_follower( $user_id, $actor );
+		}
+
+		// Handle "Undo" requests for "Like" and "Create" activities
+		if ( in_array( $type, array( 'Like', 'Create', 'Announce' ), true ) ) {
+			if ( ACTIVITYPUB_DISABLE_INCOMING_INTERACTIONS ) {
+				return;
+			}
+
+			$object_id = object_to_uri( $activity['object'] );
+			$comment   = Comment::object_id_to_comment( esc_url_raw( $object_id ) );
+
+			if ( empty( $comment ) ) {
+				return;
+			}
+
+			$state = wp_trash_comment( $comment );
+
+			do_action( 'activitypub_handled_undo', $activity, $user_id, isset( $state ) ? $state : null, null );
 		}
 	}
 }