+++ includes/model/class-blog.php
@@ -7,9 +7,11 @@ use WP_Error;
 use Activitypub\Signature;
 use Activitypub\Activity\Actor;
 use Activitypub\Collection\Users;
+use Activitypub\Collection\Extra_Fields;
 
 use function Activitypub\esc_hashtag;
 use function Activitypub\is_single_user;
+use function Activitypub\is_blog_public;
 use function Activitypub\is_user_disabled;
 use function Activitypub\get_rest_url_by_path;
 
@@ -380,7 +382,7 @@ class Blog extends Actor {
 	}
 
 	public function get_indexable() {
-		if ( \get_option( 'blog_public', 1 ) ) {
+		if ( is_blog_public() ) {
 			return true;
 		} else {
 			return false;
@@ -422,31 +424,7 @@ class Blog extends Actor {
 	 * @return array The extended User-Output.
 	 */
 	public function get_attachment() {
-		$array = array();
-
-		$array[] = array(
-			'type' => 'PropertyValue',
-			'name' => \__( 'Blog', 'activitypub' ),
-			'value' => \html_entity_decode(
-				sprintf(
-					'<a rel="me" title="%s" target="_blank" href="%s">%s</a>',
-					\esc_attr( \home_url( '/' ) ),
-					\esc_url( \home_url( '/' ) ),
-					\wp_parse_url( \home_url( '/' ), \PHP_URL_HOST )
-				),
-				\ENT_QUOTES,
-				'UTF-8'
-			),
-		);
-
-		// Add support for FEP-fb2a, for more information see FEDERATION.md
-		$array[] = array(
-			'type' => 'Link',
-			'name' => \__( 'Blog', 'activitypub' ),
-			'href' => \esc_url( \home_url( '/' ) ),
-			'rel'  => array( 'me' ),
-		);
-
-		return $array;
+		$extra_fields = Extra_Fields::get_actor_fields( $this->_id );
+		return Extra_Fields::fields_to_attachments( $extra_fields );
 	}
 }